+++ modules/documentation/docs/changelog.md
@@ -2,9 +2,12 @@
 
 ## Next Release
 
+- \[contracts\] Add `drip` task to hardhat
+
 ## 0.1.8-beta.3
 
 - \[messaging\] Fix for loop in message retry code
+
 ## 0.1.8-beta.2
 
 - \[messaging\] Fix for loop in message retry code